Version 7.0.6
Major changes to how flags are handled.  Now instead of every field modifying the same global Static FieldSettings object the FieldSettings object doesn't store the flags.
Flags are stored/loaded/read from each field directly.  This was in essence what we were doing anyways, but now it's consistent.
Changed the FieldSettings to have a DefaultFlags get/set instead of Flags.  This allows us to have the naming convention of "Default" for any flag which is read from the settings file, and "Flag" for any field-specific settings

Imported the stringtree library directly.  Was having issues including it.