Fix t: prefix doing nothing
The code previously checked if the player was in a team on a new, empty
scoreboard, which is never the case. The code has been changed to check
the main scoreboard instead.

In addition, the plugin stores allowed names as lowercase. This means
that the team check has to be case insensitive. This was previously not
the case, so this has been changed too.